Crate json_request

json_request exports a single function, request for making JSON calls to an HTTP server.

This crate relies on hyper for proving an HTTP client, and rustc_serialize for automatic encoding/decoding of JSON requests/responses.

Enums

Error
Error wrapper

Method
The Request Method (VERB)

Functions

request
Make an HTTP request

Type Definitions

Result

Implementations